used CPU time reduced with WInterHill client
I observed over 30% CPU usage with WinterHill even if no stream is played.
This patch reduced it down to 2%.
Not tested with Longmynd but shall work.
Tested with Minitiouner Express but a last verification need to be done.
While tesing I oserved a very few times  Dequeue and Peek exceptions.
The code is hacked to prevent this.